refactor(agent): extract proxy / sandbox / system_info modules
- proxy.rs: Docker (vsock→unix) and Kubernetes (vsock→TCP) API proxies
  with their per-connection bidirectional copy loops.
- sandbox.rs: SandboxService singleton + handle_sandbox_message dispatch
  + send_sandbox_error helper.
- system_info.rs: handle_get_system_info + collect_system_info reading
  /proc/{cpuinfo,meminfo,loadavg,uptime} and parsing `hostname -I`.

linux/mod.rs drops the now-unused tokio_vsock and tokio::net imports
plus the SandboxService re-export, and gains the three new submodules.
The stale "Sandbox service singleton" divider left after the move is
removed.

Pure mechanical extraction. cargo check/clippy yield the same warning
set as before.
